Kiosk/MicroControllers/led_service/requirements.txt
# MediSync RPi LED Status Service - Python dependencies
# Install: pip install -r requirements.txt

# Socket.IO async client (AIOHTTP transport support)
python-socketio[client]==5.*
aiohttp==3.*

# Raspberry Pi GPIO (PWM LED support)
gpiozero==2.*
RPi.GPIO
